Refector classes from dbUtil for biobanks in separate class
